pyproject.toml
[project]
name = "vertex-ai-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.0",
    "langchain-google-vertexai>=3.2.2",
    "python-dotenv>=1.2.1",
]
